requirements.txt
langchain
python-dotenv
ipykernel
langchain-community
pypdf
bs4
arxiv
pymupdf
wikipedia
langchain-text-splitters
langchain-openai
chromadb
sentence_transformers
langchain_huggingface
faiss-cpu
langchain_chroma
duckdb
pandas
openai
langchain-groq
duckduckgo-search
validators==0.28.1
unstructured
pytube
youtube_transcript_api
faiss-cpu
unstructured
pytube
numexpr
huggingface_hub
neo4j==5.14
openai
langchain
streamlit
python-dotenv
PyPDF2
